Sandboxes/index-factory/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.18
aiofiles==24.1.0
httpx==0.28.1
celery[redis]==5.4.0
redis==5.2.1
qdrant-client==1.12.1
open-clip-torch==2.29.0
sentence-transformers==3.3.1
Pillow==11.0.0
PyPDF2==3.0.1
beautifulsoup4==4.12.3
markdown==3.7
tiktoken==0.8.0
structlog==24.4.0
email-validator==2.2.0
